% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{adpative_function}
\alias{adpative_function}
\title{Adaptive Function for Updating Variance Components in MH and MALA Algorithms}
\usage{
adpative_function(
  index_MCMC_iter,
  sigma2_adapt,
  target_accept,
  rate_adapt,
  burn_in1,
  burn_in2,
  adapt,
  adpat_param,
  adapt_seq,
  lower.acc,
  upper.acc
)
}
\arguments{
\item{index_MCMC_iter}{Integer, the current iteration index of the MCMC algorithm.}

\item{sigma2_adapt}{Numeric, the tuning parameter for variance updates in MH and MALA algorithms.}

\item{target_accept}{Numeric, the target acceptance probability. Typically 0.224 for random walk MH and 0.57 for MALA.}

\item{rate_adapt}{Numeric, the observed acceptance rate over a fixed number of MCMC iterations.}

\item{burn_in1}{Integer, the first burn-in period for the MCMC algorithm.}

\item{burn_in2}{Integer, the second burn-in period for the MCMC algorithm.}

\item{adapt}{Integer, the sequence interval at which `sigma2_adapt` is updated.}

\item{adpat_param}{Numeric, the adaptation parameter for scaling updates to `sigma2_adapt`.}

\item{adapt_seq}{Integer vector, specifies the iterations at which adaptation occurs.}

\item{lower.acc}{Numeric, the lower bound of the acceptance rate range. For MH, it is 0.15; for MALA, it is 0.50.}

\item{upper.acc}{Numeric, the upper bound of the acceptance rate range. For MH, it is 0.30; for MALA, it is 0.65.}
}
\value{
A numeric value representing the updated `sigma2_adapt` tuning parameter.
}
\description{
This function adapts the tuning parameter for variance updates in the Metropolis-Hastings (MH) and Metropolis-Adjusted Langevin Algorithm (MALA) during MCMC iterations.
The adaptation is based on the acceptance rate of the MCMC algorithm.
}
\details{
This adaptive mechanism works by adjusting the `sigma2_adapt` parameter to keep the acceptance rate of MCMC proposals within a desirable range. It adjusts more frequently during the burn-in period and less frequently afterward. If the acceptance rate deviates too far from the target range, the function adjusts the variance to improve mixing.
}
\examples{
index_MCMC_iter <- 100
sigma2_adapt <- 1.0
target_accept <- 0.57
rate_adapt <- 0.50
burn_in1 <- 500
burn_in2 <- 200
adapt <- 50
adpat_param <- 0.1
adapt_seq <- seq(50, 1000, by = 50)
lower.acc <- 0.50
upper.acc <- 0.65
adpative_function(index_MCMC_iter, sigma2_adapt, target_accept, rate_adapt,
                  burn_in1, burn_in2, adapt, adpat_param, adapt_seq,
                  lower.acc, upper.acc)
}
