% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_fun.R
\name{cor.fun}
\alias{cor.fun}
\title{Matern Correlation Function for Different Choices of Smoothness Parameters}
\usage{
cor.fun(cor.type, dist.mat, kappa)
}
\arguments{
\item{cor.type}{A character string specifying the type of Matern correlation. There are 5 possible choices: "Matern0.5", "Matern1", "Matern1.5", "Matern2.5", and "MaternInf".}

\item{dist.mat}{A numeric matrix representing the distance matrix of dimensions n x n, where n is the total number of spatial locations.}

\item{kappa}{A numeric value representing the range parameter.}
}
\value{
A numeric matrix of the same dimensions as `dist.mat`, containing the computed correlations.
}
\description{
This function computes the Matern correlation for various choices of smoothness parameters based on the input correlation type.
}
\details{
- "Matern0.5" corresponds to the exponential correlation.
- "Matern1.5" and "Matern2.5" correspond to specific Matern correlations with different smoothness.
- "MaternInf" is also known as the squared exponential correlation, which is obtained as the smoothness parameter tends to infinity.
- "Matern1" includes the Bessel function for correlation computation, and a small value (1e-10) is added to avoid numerical instability.
}
\examples{
dist.mat <- matrix(runif(100), nrow = 10, ncol = 10)
kappa <- 1
cor.type <- "Matern1.5"
cor.fun(cor.type, dist.mat, kappa)
}
