% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{cormat.inv.fun}
\alias{cormat.inv.fun}
\title{Compute the Inverse and Log-Determinant of the Correlation Matrix}
\usage{
cormat.inv.fun(kappa, dist.mat, cor.type)
}
\arguments{
\item{kappa}{A numeric value representing the range parameter.}

\item{dist.mat}{A matrix representing the distance matrix between spatial locations.}

\item{cor.type}{A character string specifying the type of correlation. This is passed to the `cor.fun` function to determine the correlation structure.}
}
\value{
A list containing:
\describe{
  \item{cormat.inv}{The inverse of the correlation matrix.}
  \item{cormat.logdet}{The log-determinant of the correlation matrix.}
}
}
\description{
This function computes the inverse and log-determinant of a correlation matrix based on a given range parameter (`kappa`), a distance matrix (`dist.mat`), and a specified correlation type.
}
\details{
The function uses the Cholesky decomposition to efficiently compute the inverse and log-determinant of the correlation matrix. The correlation matrix is constructed using the `cor.fun` function, which takes into account the type of correlation and the distance matrix.
}
\examples{
kappa <- 1
dist.mat <- matrix(runif(9), 3, 3)
cor.type <- "Matern1.5"
result <- cormat.inv.fun(kappa, dist.mat, cor.type)
result$cormat.inv  # Inverse correlation matrix
result$cormat.logdet  # Log-determinant of the correlation matrix
}
