% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{cormat.inv.update.inla}
\alias{cormat.inv.update.inla}
\title{Update the Inverse Correlation Matrix in INLA Framework}
\usage{
cormat.inv.update.inla(rho, c.mat, g1.mat, g2.mat, alpha = 2, logDet = TRUE)
}
\arguments{
\item{rho}{A numeric value representing the range parameter.}

\item{c.mat}{A matrix involved in the update of the inverse correlation matrix.}

\item{g1.mat}{A matrix involved in the update of the inverse correlation matrix.}

\item{g2.mat}{A matrix involved in the update of the inverse correlation matrix.}

\item{alpha}{A numeric value representing the smoothness parameter, default is 2.}

\item{logDet}{Logical, if `TRUE`, the log-determinant of the updated correlation matrix will be computed and returned.}
}
\value{
A list containing:
\describe{
  \item{cormat.inv}{The updated inverse correlation matrix.}
  \item{cormat.logdet}{The log-determinant of the correlation matrix if `logDet = TRUE`, otherwise `NULL`.}
}
}
\description{
This function updates the inverse of the correlation matrix used in the INLA (Integrated Nested Laplace Approximation) framework based on given parameters and matrices.
}
\details{
The function calculates the inverse correlation matrix using the given range parameter (`rho`) and matrices (`c.mat`, `g1.mat`, `g2.mat`).
The log-determinant of the matrix is optionally calculated and returned when `logDet = TRUE`.
}
\examples{
rho <- 1
c.mat <- matrix(runif(9), 3, 3)
g1.mat <- matrix(runif(9), 3, 3)
g2.mat <- matrix(runif(9), 3, 3)
result <- cormat.inv.update.inla(rho, c.mat, g1.mat, g2.mat)
result$cormat.inv  # Inverse correlation matrix
result$cormat.logdet  # Log-determinant (if logDet = TRUE)
}
