% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{covmat.inv.update.inla}
\alias{covmat.inv.update.inla}
\title{Create Precision Matrix Based on the SPDE Triangulation}
\usage{
covmat.inv.update.inla(
  rho,
  mar.std = 1,
  c.mat,
  g1.mat,
  g2.mat,
  alpha = 2,
  logDet = TRUE
)
}
\arguments{
\item{rho}{A numeric value representing the range parameter, parametrized such that it represents the distance at which the correlations drop approximately to 0.13.}

\item{mar.std}{A numeric value for the marginal standard deviation. Default is 1.}

\item{c.mat}{A matrix involved in the update of the inverse covariance matrix.}

\item{g1.mat}{A matrix involved in the update of the inverse covariance matrix.}

\item{g2.mat}{A matrix involved in the update of the inverse covariance matrix.}

\item{alpha}{A numeric value representing the smoothness parameter, default is 2.}

\item{logDet}{Logical, if `TRUE`, the log-determinant of the updated covariance matrix will be computed and returned.}
}
\value{
A list containing:
\describe{
  \item{cormat.inv}{The updated inverse covariance matrix.}
  \item{cormat.logdet}{The log-determinant of the covariance matrix if `logDet = TRUE`, otherwise `NULL`.}
}
}
\description{
This function creates the precision matrix for a spatial process based on the Stochastic Partial Differential Equation (SPDE) triangulation. The matrix is constructed using the range parameter (`rho`), marginal standard deviation (`mar.std`), and several matrices (`c.mat`, `g1.mat`, `g2.mat`).
}
\details{
The precision matrix is calculated using the given range parameter (`rho`), marginal standard deviation (`mar.std`), and the matrices (`c.mat`, `g1.mat`, `g2.mat`) representing the SPDE structure. The function also provides the option to compute the log-determinant of the covariance matrix.
}
\examples{
rho <- 1
mar.std <- 1
c.mat <- matrix(runif(9), 3, 3)
g1.mat <- matrix(runif(9), 3, 3)
g2.mat <- matrix(runif(9), 3, 3)
result <- covmat.inv.update.inla(rho, mar.std, c.mat, g1.mat, g2.mat)
result$cormat.inv  # Inverse covariance matrix
result$cormat.logdet  # Log-determinant (if logDet = TRUE)
}
