% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{det_inv_cov_matrix}
\alias{det_inv_cov_matrix}
\title{Inverse and Determinant of a Matrix using Cholesky Decomposition}
\usage{
det_inv_cov_matrix(cov_mat, det = TRUE)
}
\arguments{
\item{cov_mat}{A numeric matrix representing the covariance matrix to be inverted.}

\item{det}{Logical, if `TRUE`, the determinant of the matrix will be computed and returned; if `FALSE`, only the inverse will be returned.}
}
\value{
A list containing:
\describe{
  \item{inv}{The inverse of the covariance matrix.}
  \item{detr}{The determinant of the matrix if `det = TRUE`, otherwise `NULL`.}
}
}
\description{
This function computes the inverse and determinant of a covariance matrix using Cholesky decomposition for efficiency.
}
\details{
The Cholesky decomposition is used to compute both the inverse and the log determinant efficiently.
The determinant is computed as the sum of the logarithms of the diagonal elements of the Cholesky factor, multiplied by 2.
}
\examples{
cov_mat <- matrix(c(4, 2, 2, 3), nrow = 2)
result <- det_inv_cov_matrix(cov_mat)
result$inv  # Inverse of the matrix
result$detr  # Determinant of the matrix

# If you only need the inverse
det_inv_cov_matrix(cov_mat, det = FALSE)
}
