% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{dlogjac_inv_transfo}
\alias{dlogjac_inv_transfo}
\title{Gradient of the Jacobian}
\usage{
dlogjac_inv_transfo(tpar, lb, ub)
}
\arguments{
\item{tpar}{The transformed parameter.}

\item{lb}{The lower bound of the parameter.}

\item{ub}{The upper bound of the parameter.}
}
\value{
A numeric value representing the gradient of the Jacobian for the transformed parameter.
}
\description{
This function computes the gradient of the Jacobian of the transformation, depending on the parameter bounds (`lb` and `ub`).
}
\details{
The gradient depends on the values of `lb` and `ub`:
- If both `lb` and `ub` are infinite, the gradient is 0.
- If `lb` is finite and `ub` is infinite, the gradient is 1.
- If `lb` is infinite and `ub` is finite, the gradient is 1.
- If both bounds are finite, the gradient is calculated as \(-1 + 2 x {plogis}(-tpar)\).
}
\examples{
dlogjac_inv_transfo(tpar = 0.5, lb = 0, ub = 1)
dlogjac_inv_transfo(tpar = 2, lb = -Inf, ub = Inf)
dlogjac_inv_transfo(tpar = 1, lb = 0, ub = Inf)
dlogjac_inv_transfo(tpar = 1, lb = -Inf, ub = 5)
}
