% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AADB_est.R
\name{est_AADB_simple}
\alias{est_AADB_simple}
\title{Estimate the Average Annual Daily Bike (AADB)}
\usage{
est_AADB_simple(data)
}
\arguments{
\item{data}{A numeric vector containing the data for which the AADB is to be estimated.}
}
\value{
A numeric value representing the estimated AADB.
}
\description{
estimate for the Average Annual Daily Bike (AADB).
}
\examples{
# Example usage:
data <- c(1, 2, 3, 4, 5)
est_AADB_simple(data)
}
