% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_scores.R
\name{eval_scores_median}
\alias{eval_scores_median}
\title{Compute Summary (Median) Statistics for Prediction Accuracy}
\usage{
eval_scores_median(
  pred_val,
  true_val,
  std_pred,
  sumry_type = "rmspe",
  alpha = 0.05
)
}
\arguments{
\item{pred_val}{A vector of predicted values.}

\item{true_val}{A vector of true values, same length as \code{pred_val}.}

\item{std_pred}{A vector of standard deviations of the predicted values, same length as \code{pred_val}.}

\item{sumry_type}{A string specifying the type of summary statistic to compute. 
Options include:
\itemize{
  \item \code{"rmspe"}: Root mean squared prediction error.
  \item \code{"mae"}: Mean absolute error.
  \item \code{"IS"}: Interval score.
  \item \code{"mpiw"}: Mean prediction interval width.
  \item \code{"picp"}: Prediction interval coverage probability.
  \item \code{"normlized_PI"}: Normalized prediction residuals.
  \item \code{"crps"}: Continuous ranked probability score.
  \item \code{"norm_pr"}: Normalized prediction residuals.
}}

\item{alpha}{A numeric value for the confidence level, default is 0.05 for a 95\% coverage probability.}
}
\value{
The calculated summary statistic based on the specified \code{sumry_type}.
}
\description{
This function calculates various summary statistics to evaluate the accuracy of predictions 
using different scoring rules. It computes statistics like mean squared prediction error, 
mean absolute error, continuous ranked probability score (CRPS), and more.
}
\details{
The summary statistics are used to evaluate the performance of the predictive distribution 
based on the true observations. All the scores are negatively oriented, meaning lower values 
indicate better performance. 

The different scoring methods include:
\itemize{
  \item \code{mspe or MSE}: Mean squared error, \code{S(F, y) = (y - m)^2}.
  \item \code{mae}: Mean absolute error, \code{S(F, y) = |y - m|}.
  \item \code{normlized_PI or IGN}: Ignorance score, \code{S(F, y) = (y - m)^2 / s^2 / 2 + log(s)}.
  \item \code{crps or CRPS}: Continuous ranked probability score.
  \item \code{IS}: Interval score.
  \item \code{mpiw}: Mean prediction interval width.
  \item \code{picp or Coverage}: Prediction interval coverage, \code{I(y in [m - q*s, m + q*s])}, 
        where \code{q} is the quantile corresponding to \code{1 - alpha/2}.
}

For more details on scoring rules, see Gneiting & Raftery, JASA, 2007.
}
\note{
Be cautious of using improper scores. For example, \code{(y - m)^2 + s^2} is not a proper score 
and should not be used.
}
\examples{
# Example usage of eval_scores_median
pred_val <- c(1.2, 2.3, 3.1)
true_val <- c(1.0, 2.0, 3.0)
std_pred <- c(0.1, 0.2, 0.3)
eval_scores_median(pred_val, true_val, std_pred, sumry_type = "rmspe")

}
