% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_theta_FFBS.R
\name{ffbs_thetas}
\alias{ffbs_thetas}
\title{Forward Filtering Backward Sampling (FFBS) for State Vectors (Thetas)}
\usage{
ffbs_thetas(p, mus, m0, m0.bar, G, Sigma, H, tau2, ns, nt)
}
\arguments{
\item{p}{Integer, the dimension of the state vector.}

\item{mus}{A matrix of dimension (nt x ns), the mean of the observations.}

\item{m0}{Initial mean vector of the state vector at time t = 0.}

\item{m0.bar}{Posterior mean vector of the state vector at time t = 0.}

\item{G}{A matrix of dimension (p x p), the state-evolution matrix.}

\item{Sigma}{A covariance matrix of the state vector.}

\item{H}{Observation matrix, mapping state vectors to observations.}

\item{tau2}{Observation noise variance.}

\item{ns}{Integer, number of spatial locations.}

\item{nt}{Integer, number of time points.}
}
\value{
A matrix of sampled state vectors (thetas) of dimension (nt x p).
}
\description{
This function implements the Forward Filtering Backward Sampling (FFBS) algorithm to draw samples of state vectors (thetas) from their posterior distribution.
}
\examples{
\dontrun{
p <- 2
ns <- 5
nt <- 100
mus <- matrix(runif(nt * ns), nrow = nt, ncol = ns)
m0 <- rep(0, p)
m0.bar <- rep(0, p)
G <- diag(p)
Sigma <- diag(p)
H <- matrix(runif(ns * p), nrow = ns, ncol = p)
tau2 <- 0.1
theta_samples <- ffbs_thetas(p, mus, m0, m0.bar, G, Sigma, H, tau2, ns, nt)
}
}
