% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_latent_lambda.R
\name{hess_lambda}
\alias{hess_lambda}
\title{Hessian of the Log Likelihood for Lambda with Gaussian Prior}
\usage{
hess_lambda(mean.lambda, tau2)
}
\arguments{
\item{mean.lambda}{A numeric value representing the mean of the Gaussian prior.}

\item{tau2}{A numeric value representing the variance of the Gaussian prior.}
}
\value{
A numeric value representing the Hessian of the log likelihood with respect to `lambda`.
}
\description{
This function computes the Hessian (second derivative) of the log likelihood for the latent parameter `lambda` under a Gaussian prior with mean `mean.lambda` and variance `tau2`.
}
\examples{
\dontrun{
# Example usage:
mean.lambda <- 0
tau2 <- 1
hess_lik <- hess_lambda(mean.lambda, tau2)
print(hess_lik)
}
}
