% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_data.R
\name{hess_lambda_data}
\alias{hess_lambda_data}
\title{Hessian of the Log-Likelihood with Respect to Lambda for Different GLM Families}
\usage{
hess_lambda_data(Y, nt, ns, mean.lambda, tau2, r, k, data_lik)
}
\arguments{
\item{Y}{A matrix of observed data.}

\item{nt}{Integer, number of temporal points.}

\item{ns}{Integer, number of spatial locations.}

\item{mean.lambda}{A matrix of the mean of lambda values.}

\item{tau2}{A scalar, variance parameter.}

\item{r}{A scalar, shape parameter for the Negative Binomial distribution.}

\item{k}{A scalar, precision parameter for the log-Normal distribution.}

\item{data_lik}{Character, the type of likelihood. Options include `"Poisson"`, `"NegB"`, and `"lNormal"`.}
}
\value{
A matrix representing the Hessian of the log-likelihood.
}
\description{
This function computes the Hessian of the log-likelihood with respect to lambda for different types of generalized linear model (GLM) families such as Poisson, Negative Binomial (NegB), and log-Normal (lNormal).
}
\examples{
\dontrun{
Y <- matrix(rpois(100, lambda=2), nrow=10, ncol=10)
nt <- 10
ns <- 10
mean.lambda <- matrix(log(2), nrow=10, ncol=10)
tau2 <- 0.1
r <- 1
k <- 0.5
hessian <- hess_lambda_data(Y, nt, ns, mean.lambda, tau2, r, k, "Poisson")
}
}
