% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_values.R
\name{init_fun}
\alias{init_fun}
\title{Initial Values for Model Parameters}
\usage{
init_fun(p, q, ns, nt, nb, Y, model, data_lik, delta, seed)
}
\arguments{
\item{p}{Integer, total number of dynamic temporal covariates.}

\item{q}{Integer, total number of fixed covariates.}

\item{ns}{Integer, number of spatial locations.}

\item{nt}{Integer, number of time points.}

\item{nb}{Integer, number of mesh nodes in the case of the SPDE approach.}

\item{Y}{A matrix representing the original data.}

\item{model}{A character string specifying the model type. The possible values are `"bayes.reg"`, `"dense"`, or `"sparse"`.}

\item{data_lik}{A character string specifying the data likelihood. Options include `"NegB"` for negative binomial or `"lNormal"` for log-normal.}

\item{delta}{A numeric value representing the maximum distance between the observed locations.}

\item{seed}{An integer seed for reproducibility.}
}
\value{
A list containing:
\describe{
  \item{init}{A list of initial parameter values based on the model type.}
  \item{param.name}{A vector of parameter names for the model.}
}
}
\description{
This function generates initial values for all the parameters of the models specified. The initial values are generated based on the provided data, model type, and likelihood. It supports three model types: `bayes.reg`, `dense`, and `sparse`.
}
\details{
This function initializes the parameters for three different models: `bayes.reg` (Bayesian regression), `dense` (dense model with space-time varying intercepts), and `sparse` (sparse model with space-time varying intercepts). The initialization process adjusts for the likelihood type and the model specifications, setting up parameters like spatial and temporal covariates, intercepts, and other components based on the data.
}
\examples{
p <- 3
q <- 2
ns <- 5
nt <- 10
nb <- 4
Y <- matrix(rnorm(50), nrow = 10, ncol = 5)
model <- "bayes.reg"
data_lik <- "lNormal"
delta <- 1
seed <- 123
result <- init_fun(p, q, ns, nt, nb, Y, model, data_lik, delta, seed)
print(result$init)
print(result$param.name)
}
