% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{inv_transfo}
\alias{inv_transfo}
\title{Transform Parameters Back to Original Scales}
\usage{
inv_transfo(tpar, lb, ub)
}
\arguments{
\item{tpar}{A numeric value representing the parameter on the unconstrained scale.}

\item{lb}{A numeric value representing the lower bound of the parameter's original scale.}

\item{ub}{A numeric value representing the upper bound of the parameter's original scale.}
}
\value{
A numeric value representing the transformed parameter back on the original scale.
}
\description{
This function transforms a parameter from an unconstrained scale (used in optimization or MCMC) back to its original scale, based on specified lower (`lb`) and upper (`ub`) bounds.
}
\details{
The transformation logic depends on the specified bounds:
- If both `lb` and `ub` are infinite, the parameter is unchanged.
- If `lb` is finite and `ub` is infinite, an exponential transformation is applied and shifted by `lb`.
- If `lb` is infinite and `ub` is finite, the parameter is transformed by subtracting an exponential term from `ub`.
- If both `lb` and `ub` are finite, the inverse logit transformation is applied to map the parameter back to the original scale.
}
\examples{
# Case with finite bounds
inv_transfo(tpar = 0.5, lb = 0, ub = 1)

# Case with lb = -Inf and ub = Inf (no transformation)
inv_transfo(tpar = 2, lb = -Inf, ub = Inf)

# Case with lb = 0 and ub = Inf (exponential transformation)
inv_transfo(tpar = 1, lb = 0, ub = Inf)

# Case with lb = -Inf and ub = 5
inv_transfo(tpar = 1, lb = -Inf, ub = 5)
}
