% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{inverse_SMW}
\alias{inverse_SMW}
\title{Inverse of a Matrix using the Sherman-Morrison-Woodbury Formula}
\usage{
inverse_SMW(alpha, tau, ns)
}
\arguments{
\item{alpha}{A numeric scalar that represents a parameter involved in the inversion calculation.}

\item{tau}{A numeric scalar that represents the diagonal elements of the matrix to be inverted.}

\item{ns}{An integer representing the dimension of the square matrix (n x n).}
}
\value{
A matrix representing the inverse of the original matrix calculated using the SMW formula.
}
\description{
This function computes the inverse of a matrix using the Sherman-Morrison-Woodbury (SMW) formula.
}
\details{
The function calculates the inverse of a matrix of the form \( A + u v^top \) efficiently using the SMW formula.
Here, the matrix \( A \) is diagonal with elements \(tau\), and \( u \) and \( v \) are vectors of 1s.
}
\examples{
alpha <- 0.5
tau <- 1
ns <- 5
inverse_SMW(alpha, tau, ns)
}
