% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{jac_inv_transfo}
\alias{jac_inv_transfo}
\title{Jacobian of the Transformations Made in \code{transfo}}
\usage{
jac_inv_transfo(tpar, lb, ub, log = FALSE)
}
\arguments{
\item{tpar}{The transformed parameter.}

\item{lb}{The lower bound of the parameter.}

\item{ub}{The upper bound of the parameter.}

\item{log}{Logical, if `TRUE`, the log-Jacobian is returned; if `FALSE`, the Jacobian is returned.}
}
\value{
A numeric value representing the Jacobian or log-Jacobian for the transformed parameter.
}
\description{
This function computes the Jacobian (or log-Jacobian) of the parameter transformations based on specified bounds (`lb`, `ub`) and returns either the Jacobian or its logarithm depending on the `log` argument.
}
\details{
The Jacobian depends on the bounds:
- If both `lb` and `ub` are infinite, the Jacobian is `log(tpar)`.
- If `lb` is finite and `ub` is infinite, the Jacobian is `tpar`.
- If `lb` is infinite and `ub` is finite, the Jacobian is `tpar`.
- If both bounds are finite, the Jacobian is `log(ub - lb) + logit(tpar) + log(1 - logit(tpar))`.
}
\examples{
jac_inv_transfo(tpar = 0.5, lb = 0, ub = 1, log = TRUE)
jac_inv_transfo(tpar = 2, lb = -Inf, ub = Inf, log = FALSE)
jac_inv_transfo(tpar = 1, lb = 0, ub = Inf, log = FALSE)
jac_inv_transfo(tpar = 1, lb = -Inf, ub = 5, log = TRUE)
}
