% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_all-models.R
\name{MCMC.sampler.st.DGLM}
\alias{MCMC.sampler.st.DGLM}
\title{MCMC Sampler for Spatio-Temporal Dynamic Generalized Linear Models (DGLM)}
\usage{
MCMC.sampler.st.DGLM(
  N.MCMC,
  model = "dense",
  data_lik = "Poisson",
  cor.type = "Matern0.5",
  Y,
  Ft,
  G,
  X,
  loc,
  spatInt.ind,
  forcast.ind,
  samples.store,
  sparse.info.sim,
  thin = 10,
  adapt = 100,
  tun_kappa = 1e-04,
  tun_lambda = 1,
  tun_r = 0.01,
  print.result = TRUE,
  traceplot = FALSE,
  true.values = NULL,
  simulation = FALSE,
  no_parallel_chain = 1
)
}
\arguments{
\item{N.MCMC}{Integer, the number of MCMC iterations.}

\item{model}{Character, specifying the model type. Possible values are `"bayes.reg"`, `"dense"`, or `"sparse"`.}

\item{data_lik}{Character, specifying the likelihood type (e.g., `"Poisson"`, `"NegB"`).}

\item{cor.type}{Character, specifying the correlation type for the dense model. Options include `"Matern0.5"`, `"Matern1"`, `"Matern1.5"`, `"Matern2.5"`, and `"MaternInf"`.}

\item{Y}{Matrix of response variables (observed and/or predicted).}

\item{Ft}{Matrix of temporal covariates.}

\item{G}{Transition matrix for the state-space model.}

\item{X}{Array of covariates (design matrix).}

\item{loc}{Matrix of spatial locations.}

\item{spatInt.ind}{Vector of indices for spatial interpolation.}

\item{forcast.ind}{Vector of indices for forecasting.}

\item{samples.store}{Integer, the number of samples to store.}

\item{sparse.info.sim}{List containing precomputed sparse model information (optional).}

\item{thin}{Integer, the thinning interval for MCMC samples.}

\item{adapt}{Integer, the adaptation interval for tuning.}

\item{tun_kappa}{Numeric, the tuning parameter for `kappa`.}

\item{tun_lambda}{Numeric, the tuning parameter for `lambda`.}

\item{tun_r}{Numeric, the tuning parameter for the `r` parameter in the negative binomial distribution.}

\item{print.result}{Logical, if `TRUE`, results will be printed.}

\item{traceplot}{Logical, if `TRUE`, trace plots of MCMC samples will be generated.}

\item{true.values}{List of true parameter values (optional).}

\item{simulation}{Logical, if `TRUE`, the function runs in simulation mode.}

\item{no_parallel_chain}{Integer, specifying the number of parallel chains (1 to 4).}
}
\value{
A list containing the summary of hyperparameters, fixed effects, dynamic temporal coefficients, space-time predictions, and trace plots of MCMC samples.
}
\description{
This function performs MCMC sampling for spatio-temporal dynamic generalized linear models (DGLM), including dense and sparse models, with support for parallel chains and space-time predictions.
}
\examples{
N.MCMC <- 1000
model <- "dense"
cor.type <- "Matern1.5"
no_parallel_chain <- 2
samples.store <- 100
thin <- 10
adapt <- 100
tun_kappa <- 1e-4
tun_lambda <- 1
Y <- matrix(rnorm(100), nrow = 10)
Ft <- matrix(rnorm(30), nrow = 10)
G <- diag(3)
X <- array(rnorm(1000), dim = c(10, 10, 3))
loc <- matrix(runif(20), ncol = 2)
spatInt.ind <- 1:2
forcast.ind <- 1:2
result <- MCMC.sampler.st.DGLM(N.MCMC, model, cor.type, no_parallel_chain,
                               samples.store, thin, adapt, tun_kappa, tun_lambda,
                               Y = Y, Ft = Ft, G = G, X = X, loc = loc,
                               spatInt.ind = spatInt.ind, forcast.ind = forcast.ind)
}
