% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space-time-prediction.R
\name{pred.dense}
\alias{pred.dense}
\title{Spatio-Temporal Prediction Based on the Fitted DGLM Full Model with Dense Correlation Structures}
\usage{
pred.dense(
  nt,
  ns,
  data_lik,
  spatInt.ind,
  forcast.ind,
  cor.type,
  pred_type = c("spatInt", "forecast", "spatInt_forecast"),
  X.intpl,
  X.frcast,
  X.st,
  Ft.o,
  Ft.frcast,
  G,
  dist.mat.all,
  kappa.samples,
  sigma2.samples,
  tau2.samples,
  theta.samples,
  Wt.samples,
  beta.samples,
  mus.samples,
  r.samples,
  k.samples
)
}
\arguments{
\item{nt}{Integer, total number of temporal points at which the model is fitted.}

\item{ns}{Integer, number of spatial locations where the data is observed and the model is fitted.}

\item{data_lik}{Character, likelihood type. Options include `"Poisson"`, `"NegB"`, and `"lNormal"`.}

\item{spatInt.ind}{Integer vector, indices of new locations for spatial interpolation or future forecasting.}

\item{forcast.ind}{Integer vector, time indices for spatial predictions (e.g., `1` means 1-step ahead prediction; `1:7` means prediction for the next 7 time points).}

\item{cor.type}{Character, specifying the correlation type for the dense model. Options include `"Matern0.5"`, `"Matern1"`, `"Matern1.5"`, `"Matern2.5"`, and `"MaternInf"`.}

\item{pred_type}{Character, type of prediction. Options are `"spatInt"` for spatial prediction at new spatial locations, `"forecast"` for future forecasting at observed sites, and `"spatInt_forecast"` for future forecasting at new spatial locations.}

\item{X.intpl}{A matrix of covariates for interpolation.}

\item{X.frcast}{A matrix of covariates for forecasting.}

\item{X.st}{A matrix of covariates for spatio-temporal predictions.}

\item{Ft.o}{A matrix of temporal covariates at observed time points.}

\item{Ft.frcast}{A matrix of temporal covariates for future forecasting.}

\item{G}{A matrix of dimension `p x p`, state-evolving matrix.}

\item{dist.mat.all}{A distance matrix of the spatial locations.}

\item{kappa.samples}{A scalar, posterior samples of the range parameter.}

\item{sigma2.samples}{A scalar, posterior samples of the variance of the spatial covariance function.}

\item{tau2.samples}{A scalar, posterior samples of the variance of the nugget term.}

\item{theta.samples}{A matrix of dimension `nt x p`, posterior samples of state vectors.}

\item{Wt.samples}{A vector of length `p`, posterior samples of the variance of state vectors.}

\item{beta.samples}{A vector, posterior samples of covariate coefficients corresponding to design matrix `X`.}

\item{mus.samples}{A matrix of dimension `nt x ns`, posterior samples of spatially varying intercept parameters.}

\item{r.samples}{A scalar, posterior samples of the shape parameter for the negative binomial distribution.}

\item{k.samples}{A scalar, posterior samples of the precision parameter for the log-normal distribution.}
}
\value{
A list containing predicted `y.pred` and `lambda.samples`.
}
\description{
This function performs spatio-temporal prediction based on the fitted DGLM full model with dense correlation structures. It supports three types of predictions: spatial interpolation at new locations, forecasting at observed sites, and forecasting at new spatial locations.
}
\examples{
\dontrun{
# Example usage:
nt <- 100
ns <- 50
beta.samples <- runif(5)
kappa.samples <- 0.5
sigma2.samples <- 0.2
tau2.samples <- 0.1
theta.samples <- matrix(runif(nt * 3), ncol = 3)
pred.dense(nt, ns, "Poisson", 1:10, 1:5, "forecast", X.intpl, X.frcast, X.st, Ft, Ft, G, dist.mat, kappa.samples, sigma2.samples, tau2.samples, theta.samples, Wt.samples, beta.samples, mus.samples)
}
}
