% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space-time-prediction.R
\name{pred.reg}
\alias{pred.reg}
\title{Spatio-Temporal Prediction Based on the Baseline Model}
\usage{
pred.reg(
  nt,
  ns,
  data_lik,
  spatInt.ind,
  forcast.ind,
  pred_type = c("spatInt", "forecast", "spatInt_forecast"),
  X.intpl,
  X.frcast,
  X.st,
  beta0.samples,
  tau2.samples,
  beta.samples,
  r.samples,
  k.samples
)
}
\arguments{
\item{nt}{Integer, total number of temporal points at which we fit the model.}

\item{ns}{Integer, number of spatial locations where the data is observed and the model is fit.}

\item{data_lik}{Character, likelihood type. Options include `"Poisson"`, `"NegB"`, and `"lNormal"`.}

\item{spatInt.ind}{Integer vector, indices of the new locations for spatial interpolation or future forecasting.}

\item{forcast.ind}{Integer vector, time indices for spatial predictions (e.g., `1` means 1-step ahead prediction; `1:7` means prediction for the next 7 time points).}

\item{pred_type}{Character, types of prediction. Options are `"spatInt"` for spatial prediction at new spatial locations, `"forecast"` for future forecasting at observed sites, and `"spatInt_forecast"` for future forecasting at new spatial locations.}

\item{X.intpl}{A matrix of covariates for interpolation.}

\item{X.frcast}{A matrix of covariates for forecasting.}

\item{X.st}{A matrix of covariates for spatio-temporal predictions.}

\item{beta0.samples}{A scalar, simulated samples of the overall intercept at a given iteration.}

\item{tau2.samples}{A scalar, simulated samples of the variance of the nugget.}

\item{beta.samples}{A vector of simulated covariate coefficients.}

\item{r.samples}{A scalar, simulated shape parameter for the negative binomial distribution.}

\item{k.samples}{A scalar, simulated precision parameter for the log-normal distribution.}
}
\value{
A list containing predicted `y.pred.samples` and `lambda.pred.samples`.
}
\description{
This function performs spatio-temporal prediction based on the baseline model for different types of predictions: spatial interpolation at new locations, forecasting at observed sites, or forecasting at new spatial locations.
}
\examples{
\dontrun{
# Example usage:
nt <- 100
ns <- 50
Xs <- matrix(runif((ns + 10) * 5), ncol = 5)
Xt <- matrix(runif((nt + 5) * 4), ncol = 4)
beta0.samples <- 0.5
tau2.samples <- 0.1
beta.samples <- runif(5)
pred_type <- "forecast"
y_pred <- pred.reg(nt, ns, "Poisson", 1:10, 1:5, pred_type, Xs, Xt, Xs, beta0.samples, tau2.samples, beta.samples, NULL, NULL)
}
}
