% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space-time-prediction.R
\name{pred.sparse}
\alias{pred.sparse}
\title{Spatio-Temporal Prediction Based on the Fitted Sparse SPDE DGLM Model}
\usage{
pred.sparse(
  nt,
  ns,
  nb,
  data_lik,
  spatInt.ind,
  forcast.ind,
  pred_type = c("spatInt", "forecast", "spatInt_forecast"),
  X.intpl,
  X.frcast,
  X.st,
  Ft.o,
  Ft.frcast,
  G,
  A.proj.o,
  A.proj.p,
  c.mat,
  g1.mat,
  g2.mat,
  kappa.samples,
  sigma2.samples,
  tau2.samples,
  theta.samples,
  Wt.samples,
  beta.samples,
  Rs.samples,
  r.samples,
  k.samples
)
}
\arguments{
\item{nt}{Integer, total number of temporal points at which the model is fitted.}

\item{ns}{Integer, number of spatial locations where the data is observed and the model is fitted.}

\item{nb}{Integer, number of mesh nodes.}

\item{spatInt.ind}{Integer vector, indices of new locations for spatial interpolation or future forecasting.}

\item{forcast.ind}{Integer vector, time indices for spatial predictions (e.g., `1` means 1-step ahead prediction; `1:7` means prediction for the next 7 time points).}

\item{pred_type}{Character, type of prediction. Options are `"spatInt"` for spatial prediction at new spatial locations, `"forecast"` for future forecasting at observed sites, and `"spatInt_forecast"` for future forecasting at new spatial locations.}

\item{X.intpl}{Numeric matrix, covariate values for spatial interpolation.}

\item{X.frcast}{Numeric matrix, covariate values for future time points.}

\item{X.st}{Numeric matrix, covariate values for both space and time.}

\item{Ft.o}{Numeric matrix, temporal covariates at observed times.}

\item{Ft.frcast}{Numeric matrix, temporal covariates for future forecasting.}

\item{G}{Matrix of dimension `p x p`, state-evolving matrix.}

\item{A.proj.o}{Numeric matrix of dimension `ns x nb`, projection matrix for observed locations.}

\item{A.proj.p}{Numeric matrix of dimension `length(spatInt.ind) x nb`, projection matrix for new spatial locations.}

\item{c.mat}{Numeric matrix, INLA mesh components.}

\item{g1.mat}{Numeric matrix, INLA mesh components.}

\item{g2.mat}{Numeric matrix, INLA mesh components.}

\item{kappa.samples}{Numeric, posterior samples of the range parameter.}

\item{sigma2.samples}{Numeric, posterior samples of the variance of the spatial covariance function.}

\item{tau2.samples}{Numeric, posterior samples of the variance of the nugget term.}

\item{theta.samples}{Matrix of dimension `nt x p`, posterior samples of state vectors.}

\item{Wt.samples}{Numeric vector of length `p`, posterior samples of the variance of state vectors.}

\item{beta.samples}{Numeric vector, posterior samples of the covariate coefficients corresponding to design matrix `X`.}

\item{Rs.samples}{Numeric vector of length `nb`, posterior samples of SPDE basis weights.}

\item{r.samples}{Numeric, posterior samples of the shape parameter for the negative binomial distribution.}

\item{k.samples}{Numeric, posterior samples of the precision parameter for the log-normal distribution.}
}
\value{
A list containing predicted `y.pred` and `lambda.samples`.
}
\description{
This function performs spatio-temporal predictions using the fitted sparse SPDE DGLM model. It supports three types of predictions: spatial interpolation at new locations, forecasting at observed sites, and forecasting at new spatial locations.
}
\examples{
\dontrun{
# Example usage:
pred.sparse(nt = 100, ns = 50, nb = 30, spatInt.ind = 1:10, forcast.ind = 1:5,
            pred_type = "forecast", X.intpl = X.intpl, X.frcast = X.frcast,
            X.st = X.st, Ft.o = Ft.o, Ft.frcast = Ft.frcast, G = G, A.proj.o = A.proj.o,
            A.proj.p = A.proj.p, c.mat = c.mat, g1.mat = g1.mat, g2.mat = g2.mat,
            kappa.samples = kappa.samples, sigma2.samples = sigma2.samples,
            tau2.samples = tau2.samples, theta.samples = theta.samples,
            Wt.samples = Wt.samples, beta.samples = beta.samples, Rs.samples = Rs.samples)
}
}
