% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{rMVNormP}
\alias{rMVNormP}
\title{Generate Samples from a Multivariate Normal Distribution with a Given Precision Matrix}
\usage{
rMVNormP(n, mu, Omega)
}
\arguments{
\item{n}{Integer, the number of samples to draw.}

\item{mu}{A numeric vector of length `p`, representing the mean of the multivariate normal distribution.}

\item{Omega}{A numeric matrix of dimension `p x p`, representing the precision matrix (the inverse of the covariance matrix).}
}
\value{
A numeric matrix of dimension `p x n`, where each column represents a sample from the multivariate normal distribution.
}
\description{
This function generates samples from a multivariate normal distribution given a precision matrix.
}
\details{
The function generates `n` samples from a multivariate normal distribution using the Cholesky decomposition of the precision matrix (`Omega`).
It first generates standard normal random variables, then transforms them using the Cholesky factor of `Omega`, and finally adds the mean vector (`mu`) to obtain the desired samples.
}
\examples{
# Example with a 3-dimensional multivariate normal distribution
mu <- c(0, 0, 0)
Omega <- diag(3)  # Precision matrix is the identity matrix
samples <- rMVNormP(n = 5, mu = mu, Omega = Omega)
print(samples)
}
