% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{transfo}
\alias{transfo}
\title{Transform Parameter to Unconstrained Scale}
\usage{
transfo(par, lb, ub)
}
\arguments{
\item{par}{A numeric scalar representing the parameter to be transformed.}

\item{lb}{A numeric scalar representing the lower bound of the parameter.}

\item{ub}{A numeric scalar representing the upper bound of the parameter.}
}
\value{
A numeric value representing the transformed parameter on the unconstrained scale.
}
\description{
This function transforms a scalar parameter from a constrained scale (bounded by `lb` and `ub`) to an unconstrained scale, which is useful in optimization and MCMC algorithms.
}
\details{
The transformation depends on the bounds (`lb`, `ub`):
- If both `lb` and `ub` are infinite, the parameter is returned unchanged.
- If `lb` is finite and `ub` is infinite, the log transformation is applied: `log(par - lb)`.
- If `lb` is infinite and `ub` is finite, the transformation is `log(ub - par)`.
- If both bounds are finite, the transformation is the logit function: `qlogis((par - lb) / (ub - lb))`.
}
\examples{
transfo(par = 0.5, lb = 0, ub = 1)
transfo(par = 2, lb = -Inf, ub = Inf)
transfo(par = 1, lb = 0, ub = Inf)
transfo(par = 1, lb = -Inf, ub = 5)
}
\references{
Section 56 of the Stan Reference manual version 2.9 at \url{https://github.com/stan-dev/stan/releases/download/v2.9.0/stan-reference-2.9.0.pdf}
}
