% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_variances.R
\name{update.Wt}
\alias{update.Wt}
\title{Update Variance of State Vectors (Wt) Using Inverse Gamma Distribution}
\usage{
\method{update}{Wt}(nt, p, theta0, theta, G, param.hyperprior)
}
\arguments{
\item{nt}{Integer, number of time points.}

\item{p}{Integer, the dimension of the state vector.}

\item{theta0}{A vector of length `p`, the initial state vector.}

\item{theta}{A matrix of dimension `nt x p`, the current state vectors at each time point.}

\item{G}{A matrix of dimension `p x p`, the state transition matrix.}

\item{param.hyperprior}{A vector of length 2, hyperprior parameters for the inverse gamma distribution (shape and rate).}
}
\value{
A vector of length `p`, simulated values of the variance for each state vector component from the inverse gamma distribution.
}
\description{
This function updates the variance of state vectors (Wt) in a dynamic linear model by drawing from an inverse gamma posterior distribution, based on the differences between the current and previous state vectors.
}
\examples{
\dontrun{
nt <- 100
p <- 5
theta0 <- runif(p)
theta <- matrix(runif(nt * p), nt, p)
G <- diag(p)
param.hyperprior <- c(0.01, 0.01)
Wt <- update.Wt(nt, p, theta0, theta, G, param.hyperprior)
}
}
