% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_intercepts.R
\name{update_intercept.dense}
\alias{update_intercept.dense}
\title{Update the Intercept for the Dense Correlation Model}
\usage{
update_intercept.dense(
  nt,
  ns,
  mu_current,
  mu0,
  lambdas_mean,
  sigma.kappa.inv,
  sigma2,
  tau2
)
}
\arguments{
\item{nt}{Integer, number of temporal points.}

\item{ns}{Integer, number of spatial locations.}

\item{mu_current}{Matrix of current intercept values with dimensions `nt x ns`.}

\item{mu0}{Vector, initial intercept for the first time step.}

\item{lambdas_mean}{Matrix of mean lambda values with dimensions `nt x ns`.}

\item{sigma.kappa.inv}{Inverse of the spatial covariance matrix.}

\item{sigma2}{Scalar, variance of the spatial covariance function.}

\item{tau2}{Scalar, variance parameter for the noise in the model.}
}
\value{
A matrix of updated intercept values of dimensions `nt x ns`.
}
\description{
This function updates the intercept in a dense correlation model using spatio-temporal data. The update is based on posterior sampling from a Gaussian distribution, accounting for the current and neighboring intercept values.
}
\examples{
\dontrun{
nt <- 10
ns <- 5
mu_current <- matrix(rnorm(nt * ns), nrow = nt, ncol = ns)
mu0 <- rep(0, ns)
lambdas_mean <- matrix(rnorm(nt * ns), nrow = nt, ncol = ns)
sigma.kappa.inv <- diag(ns)
sigma2 <- 1
tau2 <- 0.5
updated_mu <- update_intercept.dense(nt, ns, mu_current, mu0, lambdas_mean, sigma.kappa.inv, sigma2, tau2)
}
}
