% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_intercepts.R
\name{update_intercept.sparse}
\alias{update_intercept.sparse}
\title{Update Intercept for the Sparse Correlation Model}
\usage{
update_intercept.sparse(
  nt,
  ns,
  nb,
  mu_current,
  R0,
  lambdas_mean,
  Q,
  t.A.A,
  t.A,
  sigma2,
  tau2
)
}
\arguments{
\item{nt}{Integer, number of temporal points.}

\item{ns}{Integer, number of spatial locations.}

\item{nb}{Integer, number of basis functions used in the sparse representation.}

\item{mu_current}{Matrix of current intercept values with dimensions `nt x nb`.}

\item{R0}{Vector, initial intercept for the first time step.}

\item{lambdas_mean}{Matrix of mean lambda values with dimensions `nt x ns`.}

\item{Q}{Matrix, precision matrix for the sparse basis functions.}

\item{t.A.A}{Matrix, transpose of projection matrix times projection matrix.}

\item{t.A}{Matrix, transpose of projection matrix.}

\item{sigma2}{Scalar, variance of the spatial covariance function.}

\item{tau2}{Scalar, variance parameter for the noise in the model.}
}
\value{
A matrix of updated intercept values with dimensions `nt x nb`.
}
\description{
This function updates the intercept in a sparse correlation model using spatio-temporal data. The update is based on posterior sampling from a Gaussian distribution, incorporating both current and neighboring intercept values in the sparse matrix structure.
}
\examples{
\dontrun{
nt <- 10
ns <- 5
nb <- 3
mu_current <- matrix(rnorm(nt * nb), nrow = nt, ncol = nb)
R0 <- rep(0, nb)
lambdas_mean <- matrix(rnorm(nt * ns), nrow = nt, ncol = ns)
Q <- diag(nb)
t.A.A <- diag(nb)
t.A <- diag(nb)
sigma2 <- 1
tau2 <- 0.5
updated_mu <- update_intercept.sparse(nt, ns, nb, mu_current, R0, lambdas_mean, Q, t.A.A, t.A, sigma2, tau2)
}
}
