% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_intercepts.R
\name{update_intercept_reg}
\alias{update_intercept_reg}
\title{Update the Intercept for the Regression Model}
\usage{
update_intercept_reg(nt, ns, lambda.mean.diff, tau2, prior.var)
}
\arguments{
\item{nt}{Integer, number of temporal points.}

\item{ns}{Integer, number of spatial locations.}

\item{lambda.mean.diff}{A matrix of the difference between lambda and the mean values.}

\item{tau2}{A scalar, variance parameter of the likelihood.}

\item{prior.var}{A scalar, variance parameter of the prior distribution for the intercept.}
}
\value{
A scalar representing the updated intercept.
}
\description{
This function updates the intercept in a regression model using the posterior distribution, incorporating the likelihood from the observed data and a prior distribution.
}
\examples{
\dontrun{
nt <- 10
ns <- 5
lambda.mean.diff <- matrix(rnorm(nt * ns), nrow = nt, ncol = ns)
tau2 <- 0.1
prior.var <- 1
intercept <- update_intercept_reg(nt, ns, lambda.mean.diff, tau2, prior.var)
}
}
