% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_kappa.R
\name{update_kappa.dense}
\alias{update_kappa.dense}
\title{Update the Range Parameter (Kappa) for the Dense Correlation Model}
\usage{
update_kappa.dense(
  nt,
  cur.kappa,
  mu,
  mu0,
  sigma2,
  log.det,
  cov_inv,
  delta,
  tun.kappa,
  dist.mat,
  cor.type,
  ind = FALSE
)
}
\arguments{
\item{nt}{Integer, number of temporal points.}

\item{cur.kappa}{Scalar, current value of the range parameter (kappa).}

\item{mu}{Matrix of current intercept values with dimensions `nt x ns`.}

\item{mu0}{Vector, prior mean for the intercept.}

\item{sigma2}{Scalar, variance of the spatial covariance function.}

\item{log.det}{Scalar, log determinant of the covariance matrix.}

\item{cov_inv}{Matrix, inverse of the covariance matrix.}

\item{delta}{Scalar, tuning parameter for the proposal distribution.}

\item{tun.kappa}{Scalar, tuning parameter for adjusting the proposal variance.}

\item{dist.mat}{Matrix, distance matrix between spatial locations.}

\item{cor.type}{Character, type of correlation function to use (e.g., "exponential").}

\item{ind}{Logical, whether to use independent updating.}
}
\value{
Updated value of the range parameter (kappa).
}
\description{
This function updates the range parameter (kappa) in a dense correlation model using a Metropolis-Hastings step. The update is based on the current kappa, the spatio-temporal data, and covariance matrix.
}
\examples{
\dontrun{
cur.kappa <- 0.5
nt <- 100
sigma2 <- 1
log.det <- 0
delta <- 0.01
tun.kappa <- 0.1
dist.mat <- matrix(runif(25), nrow = 5)
mu <- matrix(rnorm(500), nrow = nt, ncol = 5)
mu0 <- rep(0, 5)
cov_inv <- diag(5)
cor.type <- "exponential"
updated_kappa <- update_kappa.dense(nt, cur.kappa, mu, mu0, sigma2, log.det, cov_inv, delta, tun.kappa, dist.mat, cor.type)
}
}
