% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_kappa.R
\name{update_kappa.sparse}
\alias{update_kappa.sparse}
\title{Update Kappa using Metropolis-Hastings for Sparse Model}
\usage{
update_kappa.sparse(
  nt,
  cur.kappa,
  sigma2,
  c.mat,
  g1.mat,
  g2.mat,
  R,
  R0,
  cor.mat.inv,
  log.det.cor.mat,
  delta,
  tun.kappa,
  ind = FALSE
)
}
\arguments{
\item{nt}{Integer, number of temporal points.}

\item{cur.kappa}{Scalar, current value of the range parameter (kappa).}

\item{sigma2}{Scalar, variance of the spatial covariance function.}

\item{c.mat}{Matrix, constant matrix in the SPDE precision formulation.}

\item{g1.mat}{Matrix, part of the first-order spatial derivative matrix.}

\item{g2.mat}{Matrix, part of the second-order spatial derivative matrix.}

\item{R}{Matrix, current values of the latent process.}

\item{R0}{Matrix, prior values of the latent process at time `t = 0`.}

\item{cor.mat.inv}{Matrix, inverse of the current correlation matrix.}

\item{log.det.cor.mat}{Scalar, log determinant of the correlation matrix.}

\item{delta}{Scalar, tuning parameter for the proposal distribution.}

\item{tun.kappa}{Scalar, tuning parameter for adjusting the proposal variance.}

\item{ind}{Logical, whether to use independent updating.}
}
\value{
Updated value of the range parameter (kappa).
}
\description{
This function updates the range parameter (kappa) for a sparse spatial model using a Metropolis-Hastings algorithm. It considers the sparse precision matrix structure and spatial covariance.
}
\examples{
\dontrun{
cur.kappa <- 0.5
sigma2 <- 1
c.mat <- matrix(runif(25), nrow = 5)
g1.mat <- matrix(runif(25), nrow = 5)
g2.mat <- matrix(runif(25), nrow = 5)
cor.mat.inv <- diag(5)
log.det.cor.mat <- 0
delta <- 0.01
tun.kappa <- 0.1
nt <- 100
R <- matrix(rnorm(500), nrow = nt, ncol = 5)
R0 <- rep(0, 5)
updated_kappa <- update_kappa.sparse(nt, cur.kappa, sigma2, c.mat, g1.mat, g2.mat, R, R0, cor.mat.inv, log.det.cor.mat, delta, tun.kappa)
}
}
