% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_lambda.R
\name{update_lambda_MMALA}
\alias{update_lambda_MMALA}
\title{Update Lambda Using MMALA (Manifold Metropolis-Adjusted Langevin Algorithm)}
\usage{
update_lambda_MMALA(
  ns,
  nt,
  Y,
  cur.lambda,
  mean.lambda,
  tau2,
  r,
  k,
  data_lik,
  tun_lambda
)
}
\arguments{
\item{ns}{Integer, number of spatial locations.}

\item{nt}{Integer, number of time points.}

\item{Y}{Matrix of observed data with dimensions `nt x ns`.}

\item{cur.lambda}{Matrix, current values of lambda with dimensions `nt x ns`.}

\item{mean.lambda}{Matrix, mean of lambda with dimensions `nt x ns`.}

\item{tau2}{Scalar, variance of the nugget effect.}

\item{r}{Scalar, shape parameter for the Negative Binomial distribution.}

\item{k}{Scalar, precision parameter for the Log-Normal distribution.}

\item{data_lik}{Character, likelihood type. Options are `"Poisson"`, `"NegB"`, or `"lNormal"`.}

\item{tun_lambda}{Scalar, tuning parameter for the lambda proposal variance.}
}
\value{
A list with:
  - `cur_lambda`: Updated lambda values.
  - `ind.acc`: Acceptance indicator for each location and time point.
}
\description{
This function updates the lambda parameter using the MMALA method in a spatio-temporal model. It computes the proposal for lambda based on the gradient and Hessian information for different likelihoods (Poisson, Negative Binomial, and Log-Normal).
}
\examples{
\dontrun{
ns <- 50
nt <- 100
Y <- matrix(rpois(ns * nt, lambda = 1), ncol = ns)
cur.lambda <- matrix(rnorm(ns * nt), ncol = ns)
mean.lambda <- matrix(runif(ns * nt), ncol = ns)
tau2 <- 0.5
r <- 1
k <- 1
data_lik <- "Poisson"
tun_lambda <- 0.1
result <- update_lambda_MMALA(ns, nt, Y, cur.lambda, mean.lambda, tau2, r, k, data_lik, tun_lambda)
}
}
