% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_coeff_design.R
\name{update_purely_spat_beta}
\alias{update_purely_spat_beta}
\title{Update Fixed Effects Coefficients With Purely Spatial Covariates}
\usage{
update_purely_spat_beta(ns, nt, X, sigma2, Sigma.inv, mean_lat, var_hyprior)
}
\arguments{
\item{ns}{Integer, the number of spatial locations.}

\item{nt}{Integer, the number of temporal points.}

\item{X}{A matrix of covariates of dimension `ns x p`.}

\item{sigma2}{A scalar, the variance parameter.}

\item{Sigma.inv}{A matrix, the inverse of the covariance matrix.}

\item{mean_lat}{A matrix, the mean of latent variables.}

\item{var_hyprior}{A scalar, the hyperprior variance for the beta coefficients.}
}
\value{
A matrix of updated beta coefficients.
}
\description{
This function updates the purely spatial beta coefficients using the posterior mean and covariance matrix for a spatio-temporal model.
}
\examples{
\dontrun{
ns <- 50
nt <- 100
X <- matrix(rnorm(ns * 5), nrow = ns, ncol = 5)
sigma2 <- 0.1
Sigma.inv <- diag(ns)
mean_lat <- matrix(rnorm(ns * nt), nrow = ns, ncol = nt)
var_hyprior <- 0.01
beta <- update_purely_spat_beta(ns, nt, X, sigma2, Sigma.inv, mean_lat, var_hyprior)
}
}
