% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_coeff_design.R
\name{update_purely_temp_beta}
\alias{update_purely_temp_beta}
\title{Update Fixed Effects Coefficients With Purely Temporal Covariates}
\usage{
update_purely_temp_beta(
  p,
  ns,
  X,
  sigma2,
  Sigma.inv,
  mean_lat,
  cov_quad,
  var_hyprior
)
}
\arguments{
\item{p}{Integer, the number of covariates.}

\item{ns}{Integer, the number of spatial locations.}

\item{X}{A matrix of covariates.}

\item{sigma2}{A scalar, the variance parameter.}

\item{Sigma.inv}{A matrix, the inverse of the covariance matrix.}

\item{mean_lat}{A vector, the mean of latent variables.}

\item{cov_quad}{A matrix, the quadratic form of covariates.}

\item{var_hyprior}{A scalar, the hyperprior variance for the beta coefficients.}
}
\value{
A matrix of updated beta coefficients.
}
\description{
This function updates the purely temporal beta coefficients using the posterior mean and covariance matrix.
}
\examples{
\dontrun{
p <- 5
ns <- 50
X <- matrix(rnorm(500), nrow = ns, ncol = p)
sigma2 <- 0.1
Sigma.inv <- diag(ns)
mean_lat <- rnorm(ns)
cov_quad <- t(X) \%*\% X
var_hyprior <- 0.01
beta <- update_purely_temp_beta(p, ns, X, sigma2, Sigma.inv, mean_lat, cov_quad, var_hyprior)
}
}
