% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_NB_hyper.R
\name{update_r}
\alias{update_r}
\title{Update Shape Parameter (r) for Negative Binomial Distribution}
\usage{
update_r(Y, lambda, r, tun.r, ind = FALSE)
}
\arguments{
\item{Y}{Observed data, a matrix.}

\item{lambda}{Mean parameter for the Negative Binomial distribution.}

\item{r}{Current value of the shape parameter.}

\item{tun.r}{Tuning parameter for the proposal distribution of r.}

\item{ind}{Logical, indicates whether the proposal was accepted.}
}
\value{
A list with:
  - `r`: Updated value of the shape parameter.
  - `ind`: Acceptance indicator.
}
\description{
This function updates the shape parameter (r) for the Negative Binomial distribution using the Metropolis-Hastings algorithm.
}
\examples{
\dontrun{
Y <- matrix(rnbinom(100, mu = 10, size = 5), ncol = 10)
lambda <- matrix(log(10), ncol = 10)
r <- 5
tun.r <- 0.1
result <- update_r(Y, lambda, r, tun.r)
}
}
