% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-SPDE-basis-weights.R
\name{update_spde_weights.M1_sp}
\alias{update_spde_weights.M1_sp}
\title{Updating SPDE Basis Weight Parameters Using Gibbs Sampling}
\usage{
update_spde_weights.M1_sp(nb, nt, tau2, Q, t.A.A, latent.mean.parts)
}
\arguments{
\item{nb}{Integer, the number of basis functions.}

\item{nt}{Integer, the number of time points.}

\item{tau2}{A scalar, variance of the nugget term.}

\item{Q}{A matrix of dimension `nb x nb`, the precision matrix of the SPDE basis.}

\item{t.A.A}{A matrix, crossproduct of the projection matrix that projects the SPDE basis nodes to the observed spatial locations.}

\item{latent.mean.parts}{A matrix, terms that appear in the mean of the SPDE basis weight parameters.}
}
\value{
A matrix of updated SPDE basis weight parameters.
}
\description{
This function updates the SPDE basis weight parameters using Gibbs sampling. It relies on the precision matrix `Q`, the projection matrix `A` (through its crossproduct `t.A.A`), and the latent mean terms.
}
\examples{
\dontrun{
nb <- 50
nt <- 100
tau2 <- 0.1
Q <- diag(nb)
t.A.A <- matrix(runif(nb * nb), nb, nb)
latent.mean.parts <- matrix(runif(nt * nb), nt, nb)
spde_weights <- update_spde_weights.M1_sp(nb, nt, tau2, Q, t.A.A, latent.mean.parts)
}
}
