% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_variances.R
\name{update_var}
\alias{update_var}
\title{Update Variance Hyperparameter Using Inverse Gamma Distribution}
\usage{
update_var(ns, nt, param.hyperprior, qud_sum)
}
\arguments{
\item{ns}{Integer, number of spatial locations.}

\item{nt}{Integer, number of time points.}

\item{param.hyperprior}{A vector of length 2, hyperprior parameters for the inverse gamma distribution (shape and rate).}

\item{qud_sum}{A numeric value, the sum of squared differences or residuals.}
}
\value{
A simulated variance value from the inverse gamma distribution.
}
\description{
This function updates the variance parameter by drawing from an inverse gamma posterior distribution, typically used for variance components in Bayesian hierarchical models.
}
\examples{
\dontrun{
ns <- 5
nt <- 100
param.hyperprior <- c(0.01, 0.01)
qud_sum <- 10.5
updated_variance <- update_var(ns, nt, param.hyperprior, qud_sum)
}
}
